+++ backend/src/main/java/com/now/naaga/auth/annotation/Auth.java
@@ -1,4 +1,4 @@
-package com.now.naaga.config;
+package com.now.naaga.auth.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
