+++ backend/src/main/java/com/now/naaga/auth/presentation/AuthenticationArgumentResolver.java
@@ -1,7 +1,8 @@
-package com.now.naaga.config;
+package com.now.naaga.auth.presentation;
 
-import com.now.naaga.config.infrastructure.AuthorizationExtractor;
-import com.now.naaga.member.dto.MemberRequest;
+import com.now.naaga.auth.annotation.Auth;
+import com.now.naaga.member.dto.MemberCommand;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
@@ -9,21 +10,15 @@ import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
-import static com.now.naaga.config.infrastructure.AuthorizationExtractor.AUTHORIZATION;
+import static com.now.naaga.auth.presentation.AuthInterceptor.AUTH_KEY;
 
 @Component
 public class AuthenticationArgumentResolver implements HandlerMethodArgumentResolver {
 
-    final AuthorizationExtractor<MemberRequest> authorizationExtractor;
-
-    public AuthenticationArgumentResolver(AuthorizationExtractor<MemberRequest> authorizationExtractor) {
-        this.authorizationExtractor = authorizationExtractor;
-    }
-
     @Override
     public boolean supportsParameter(final MethodParameter parameter) {
         return parameter.hasParameterAnnotation(Auth.class)
-                || parameter.getParameterType().isInstance(MemberRequest.class);
+                && parameter.getParameterType().equals(MemberCommand.class);
     }
 
     @Override
@@ -32,7 +27,8 @@ public class AuthenticationArgumentResolver implements HandlerMethodArgumentReso
             final ModelAndViewContainer mavContainer,
             final NativeWebRequest webRequest,
             final WebDataBinderFactory binderFactory) throws Exception {
-        final String header = webRequest.getHeader(AUTHORIZATION);
-        return authorizationExtractor.extract(header);
+
+        final HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
+        return request.getAttribute(AUTH_KEY);
     }
 }
